Add NetworkingContext to avoid layer violations https://bugs.webkit.org/show_bug.cgi?id=42292
Patch by Jesus Sanchez-Palencia <jesus.palencia@openbossa.org> on 2010-08-17
Reviewed by Darin Adler.

Preparation: Just add the files to the build system.

.:

* GNUmakefile.am: Added new files.

WebCore:

* GNUmakefile.am: Added new files.
* WebCore.gypi: Ditto.
* WebCore.pro: Ditto.
* WebCore.vcproj/WebCore.vcproj: Ditto.
* WebCore.xcodeproj/project.pbxproj: Ditto.

* loader/FrameNetworkingContext.h: Added.
* platform/network/NetworkingContext.h: Added.
Placeholders with tentative code that might be changed when landing
the rest of it.

WebKit:

* WebKit.xcodeproj/project.pbxproj: Added new files.

WebKit/chromium:

* WebKit.gyp: Added new files.
* src/FrameNetworkingContextImpl.h: Added.
Placeholder with tentative code that might be changed when landing
the rest of it.

WebKit/gtk:

* WebCoreSupport/FrameNetworkingContextGtk.h: Added.
Placeholder with tentative code that might be changed when landing
the rest of it.

WebKit/mac:

* WebCoreSupport/WebFrameNetworkingContext.h: Added.
Placeholder with tentative code that might be changed when landing
the rest of it.
* WebCoreSupport/WebFrameNetworkingContext.mm: Added.
Empty placeholder for now.

WebKit/qt:

* WebCoreSupport/FrameNetworkingContextQt.cpp: Added.
Empty placeholder for now.
* WebCoreSupport/FrameNetworkingContextQt.h: Added.
Placeholder with tentative code that might be changed when landing
the rest of it.

WebKit/win:

* WebCoreSupport/WebFrameNetworkingContext.cpp: Added.
Empty placeholder for now.
* WebCoreSupport/WebFrameNetworkingContext.h: Added.
Placeholder with tentative code that might be changed when landing
the rest of it.
* WebKit.vcproj/WebKit.vcproj: Added new files.

WebKit2:

* WebKit2.xcodeproj/project.pbxproj: Added new files.
* WebProcess/WebCoreSupport/mac/WebFrameNetworkingContext.h: Added.
Placeholder with tentative code that might be changed when landing
the rest of it.
* WebProcess/WebCoreSupport/mac/WebFrameNetworkingContext.mm: Added.
Empty placeholder for now.



Identifier: 56362@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@65579 268f45cc-cd09-0410-ab3c-d52691b4dbfc