Follow-up to r41020.
Reviewed by Mark Rowe.

No change in behavior, thus no new tests.

* platform/graphics/mac/FontPlatformDataMac.mm:
(WebCore::FontPlatformData::FontPlatformData): Asserted that the NSFont parameter is not nil,
and removed nil check.
(WebCore::FontPlatformData::setFont): Asserted that the NSFont parameter is not nil and that
this is not the deleted value, and removed code to handle those cases.



Identifier: 56507@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@65727 268f45cc-cd09-0410-ab3c-d52691b4dbfc