[Qt] Add support for binding QtWebKit to a QScriptEngine.
Reviewed by Andreas Kling.

The ScriptController owns the V8 context, and this patch
for the Qt build also gives it a QScriptEngine, that has
been instructed to adopt the V8 context.

This is the enabler for the QObject bindings, used in
the API and DRT.

* WebCore.pro:
* bindings/v8/ScriptController.cpp: Include QScriptEngine
for the Qt build, to let the ~ScriptController destructor
find the QScriptEngine destructor (via OwnPtr).
* bindings/v8/ScriptController.h: Declare getters.
* bindings/v8/ScriptControllerQt.cpp: Added.
(WebCore::ScriptController::qtScriptEngine): Implement
engine adoption.

Identifier: 58067@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@67335 268f45cc-cd09-0410-ab3c-d52691b4dbfc