Resize corner jiggles when resizing slowly in WebKit2 <rdar://problem/7897425> https://bugs.webkit.org/show_bug.cgi?id=45601
Reviewed by Anders Carlsson.

* WebProcess/WebCoreSupport/WebChromeClient.cpp:
(WebKit::WebChromeClient::windowResizerRect):
* WebProcess/WebPage/WebPage.cpp:
(WebKit::WebPage::windowResizerRect):
* WebProcess/WebPage/WebPage.h:
Move resize corner logic to WebPage and use the FrameView's size instead of the WebPages
stored size, which doesn't get updated at the right time. This is in line with all other
scrollbar metrics which are in terms of the FrameView's size.



Identifier: 58467@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@67761 268f45cc-cd09-0410-ab3c-d52691b4dbfc