* Platform/CoreIPC/Connection.h: (CoreIPC::Connection::sendSync): Add new sendSync overload that takes a message struct.
Reviewed by Sam Weinig.

* PluginProcess/WebProcessConnection.cpp:
(WebKit::WebProcessConnection::didReceiveSyncMessage):
Call didReceiveSyncWebProcessConnectionMessage.

* WebProcess/Plugins/PluginProcessConnection.cpp:
(WebKit::PluginProcessConnection::addPluginProxy):
Add the plug-in proxy to the map.

(WebKit::PluginProcessConnection::removePluginProxy):
Remove the plug-in proxy from the map. If the map is empty, disconnect from the
plug-in process.

* WebProcess/Plugins/PluginProxy.cpp:
(WebKit::generatePluginInstanceID):
Generate a unique plug-in instance ID.

(WebKit::PluginProxy::PluginProxy):
Add the plug-in proxy to the map.

(WebKit::PluginProxy::~PluginProxy):
Remove the plug-in proxy from the map.

(WebKit::PluginProxy::initialize):
Ask the plug-in process to create a plug-in.

Identifier: 59140@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@68452 268f45cc-cd09-0410-ab3c-d52691b4dbfc