[Qt] Fix the build on non-x86 platforms with gcc
Reviewed by Csaba Osztrogonác.

Don't disable -Werror on arm, that doesn't work when QT_ARCH is
for example armv6. Instead change the condition as the comment says,
enable -Werror on x86/gcc where we know it passes. On other platforms
gcc produces difference warnings, and when they're fixed we can add
them to the whitelist of -Werror-supported architectures.

* WebKit.pri:

Identifier: 59299@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@68637 268f45cc-cd09-0410-ab3c-d52691b4dbfc