AX: doAXRangeForLine does not work https://bugs.webkit.org/show_bug.cgi?id=47101
Reviewed by Beth Dakin.

WebCore:

Asking for NSAccessibilityRangeForLine was returning a null range for any line number > 0.
The code was using a SelectionController to extend to the next line. Rather than change the implementation
of that core functionality, it is cleaner to use endOfLine to find the end of the line.

Test: platform/mac/accessibility/range-for-line-textarea.html

* accessibility/AccessibilityRenderObject.cpp:
(WebCore::AccessibilityRenderObject::doAXRangeForLine):

WebKitTools:

DRT support to handle NSAccessibilityRangeForLineParameterizedAttribute.

* DumpRenderTree/AccessibilityUIElement.cpp:
(rangeForLineCallback):
(AccessibilityUIElement::rangeForLine):
(AccessibilityUIElement::getJSClass):
* DumpRenderTree/AccessibilityUIElement.h:
* DumpRenderTree/mac/AccessibilityUIElementMac.mm:
(AccessibilityUIElement::rangeForLine):

LayoutTests:

* platform/mac/accessibility/range-for-line-textarea-expected.txt: Added.
* platform/mac/accessibility/range-for-line-textarea.html: Added.



Identifier: 59640@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@69033 268f45cc-cd09-0410-ab3c-d52691b4dbfc