Unreviewed build fix.
[Symbian] Fix build issue due to gaps in ordinals in the def file

Restore previously removed symbols and mark them absent instead. We must
not remove symbols from this file, because it breaks binary compatibility
(when re-assigning the ordinals of the following symbols) or it causes build
breakage when instead having a gap in the ordinal sequence. Instead removing
a symbol is done by marking it as absent.

* symbian/eabi/QtWebKitu.def:

Identifier: 59711@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@69104 268f45cc-cd09-0410-ab3c-d52691b4dbfc