Start cleaning up Arguments.h https://bugs.webkit.org/show_bug.cgi?id=47304
Reviewed by Sam Weinig.

JavaScriptCore:

* wtf/TypeTraits.h:
* wtf/TypeTraits.cpp:
Add RemoveReference type trait.

WebKit2:

This is a first step towards simplifying all the handleMessage overloads we have.

* Platform/CoreIPC/Arguments.h:
(CoreIPC::Arguments1):
Make Arguments1 a struct and get rid of the FirstArgumentType typedef. Instead, add a
ValueType typedef which is defined to an Arguments1 type whose type parameter is not
a reference type.

* Platform/CoreIPC/HandleMessage.h:
(CoreIPC::handleMessage):
Update for the Arguments1 change. Rename "arguments" to "argumentDecoder" and "reply" to
"replyEncoder".

Identifier: 59839@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@69236 268f45cc-cd09-0410-ab3c-d52691b4dbfc