Teach update-webkit-support-libs about the new versioning of WebKitSupportLibrary
Fixes <http://webkit.org/b/47915> update-webkit-support-libs should
check version numbers instead of modification times

Reviewed by Sam Weinig.

* Scripts/update-webkit-support-libs: Fetch the expected version
number from developer.apple.com, then compare with the version number
of the extracted library and of the zipped library to see if anything
needs to be done. Removed code that tracked the modified timestamp of
the library, as it is no longer needed.

Identifier: 60684@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@70128 268f45cc-cd09-0410-ab3c-d52691b4dbfc