Crash evaluating JavaScript string that throws an exception https://bugs.webkit.org/show_bug.cgi?id=48092 <rdar://problem/8487657>
Reviewed by Adam Roben.

WebKit2:

* UIProcess/WebProcessProxy.cpp:
(WebKit::WebProcessProxy::sendMessage):
m_connection can be null here; just return false if that is the case.

* WebProcess/WebPage/WebPage.cpp:
(WebKit::WebPage::runJavaScriptInMainFrame):
If the JSValue returned is null we just send along a null string.

WebKitTools:

Add a test that evaluates a JavaScript string that throws an exception and check that
the callback is called.

* TestWebKitAPI/TestWebKitAPI.xcodeproj/project.pbxproj:
* TestWebKitAPI/Tests/WebKit2/EvaluateJavaScript.cpp: Added.
(TestWebKitAPI::didRunJavaScript):
(TestWebKitAPI::WebKit2_EvaluateJavaScript):
* TestWebKitAPI/win/TestWebKitAPI.vcproj:

Identifier: 60818@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@70267 268f45cc-cd09-0410-ab3c-d52691b4dbfc