Fix assert when a NetscapePluginModule loses its last NetscapePlugin. https://bugs.webkit.org/show_bug.cgi?id=48155
Reviewed by Adam Roben.

(The assert in question is ASSERT(pluginModuleIndex != notFound) in
NetscapePluginModule::shutdown()).

* WebProcess/Plugins/Netscape/NetscapePluginModule.cpp:
(WebKit::NetscapePluginModule::pluginCreated):
If we didn't have any live plug-ins, load the plug-in module.

(WebKit::NetscapePluginModule::pluginDestroyed):
If this was the last plug-in, unload the plug-in module.

(WebKit::NetscapePluginModule::shutdown):
Set m_isInitialized to false.

(WebKit::NetscapePluginModule::load):
Don't do anything if the plug-in module has already been initialized.

(WebKit::NetscapePluginModule::unload):
Assert that we're not initialized.

* WebProcess/Plugins/Netscape/mac/NetscapePluginMac.mm:
(WebKit::NetscapePlugin::platformPostInitialize):
Assert that we successfully created the window.

Identifier: 60891@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@70347 268f45cc-cd09-0410-ab3c-d52691b4dbfc