WebContext::ensureWebProcess shouldn't send multiple InitializeWebProcess messages. https://bugs.webkit.org/show_bug.cgi?id=48158 <rdar://problem/8512572>
Reviewed by Sam Weinig.

If the web process was launching when another call to ensureWebProcess is sent, we
would end up sending another InitializeWebProcess message to it which would result
in assertions firing. Fix this by only checking for m_process, and not requiring a process
to be valid (which it won't be until it's finished launching).

* UIProcess/WebContext.cpp:
(WebKit::WebContext::ensureWebProcess):

Identifier: 60899@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@70359 268f45cc-cd09-0410-ab3c-d52691b4dbfc